+++ src/dealer/dealer.ts
@@ -26,12 +26,12 @@ export class Dealer extends Role {
         const secondCard = this.hand[1];
         const hiddenCard = document.getElementById(`${secondCard.face}-hidden`) as HTMLImageElement;
 
-        if (hiddenCard) {
-            hiddenCard.src = `./assets/${secondCard.face}.svg`;
-            hiddenCard.alt = secondCard.face;
-            hiddenCard.id = hiddenCard.alt;
-            super.writeScore(this.score.toString());
-        }
+        hiddenCard.className = `flip ${hiddenCard.className}`;
+        hiddenCard.src = `./assets/${secondCard.face}.svg`;
+        hiddenCard.alt = secondCard.face;
+        hiddenCard.id = secondCard.face;
+
+        super.writeScore(this.score.toString());
     }
 
     get isInitialHand(): boolean {