+++ src/player/player.js
@@ -1,4 +1,3 @@
-import { Game as Rules } from "../game/game.constants.js";
 import { Role } from "../role/role.js";
 export class Player extends Role {
     constructor() {
@@ -7,36 +6,72 @@ export class Player extends Role {
         this.money = 100;
         this.bet = 0;
         this.moneyDisplay = document.getElementById(`${PLAYER}-money`);
-        this.refreshMoney();
+        this.betDisplay = document.getElementById(`${PLAYER}-bet`);
+        this.resultDisplay = document.getElementById('result-display');
+        this.resultType = document.getElementById('result-type');
+        this.resultMoney = document.getElementById('result-money');
+        this.refreshDisplay();
+    }
+    get payment() {
+        return this.bet;
     }
     placeBet() {
         const bet = document.getElementById('bet').value;
-        this.bet = Number.parseInt(bet);
+        this.bet = Number(bet);
         this.money -= this.bet;
-        this.refreshMoney();
+        this.refreshDisplay();
     }
     refreshMoney() {
         this.moneyDisplay.innerText = `$${this.money.toString()}`;
     }
-    refreshMoneyAfterResult(resultFunc) {
-        resultFunc.call(this);
+    refreshBet() {
+        this.betDisplay.innerText = this.bet.toString();
+    }
+    refreshDisplay(resultFunc) {
+        let moneyClass = 'refresh';
+        if (resultFunc) {
+            resultFunc.call(this);
+            moneyClass = `money-${resultFunc.name}`;
+            this.bet = 0;
+        }
+        this.betDisplay.className = 'refresh';
+        this.moneyDisplay.className = moneyClass;
         this.refreshMoney();
-        this.setResultClass(resultFunc.name);
+        this.refreshBet();
+        setTimeout(() => {
+            this.moneyDisplay.className = '';
+            this.betDisplay.className = '';
+        }, 5e2);
+    }
+    displayResult(resultFunc) {
+        const type = resultFunc.name;
+        this.resultType.innerText = type;
+        let money = '';
+        if (type === this.win.name) {
+            money = `+$${this.payment.toString()}`;
+        }
+        else if (type === this.bust.name) {
+            money = `-$${this.bet.toString()}`;
+        }
+        this.resultMoney.innerText = money;
+        this.resultDisplay.className = 'result-display';
+        this.resultType.className = `result-type-${type}`;
+        this.resultMoney.className = `result-money-${type}`;
+        setTimeout(() => {
+            this.resultType.className = '';
+            this.resultMoney.className = '';
+            this.resultDisplay.className = '';
+            this.refreshDisplay(resultFunc);
+        }, 3e3);
     }
     bust() {
         if (this.money <= 0) {
-            alert("No money left, game will restart");
+            alert("You've run out of money. \nGame will restart, good luck!");
             location.reload();
         }
     }
     win() {
-        this.money += (this.bet * Rules.ODDS);
-    }
-    setResultClass(result) {
-        this.moneyDisplay.className = `money-${result}`;
-        setTimeout(() => {
-            this.moneyDisplay.className = '';
-        }, 15e2);
+        this.money += this.payment;
     }
     /**
      * Occurs when the player and the dealer have the same total value for their hands at the end of a round.