+++ CHANGELOG.md
@@ -5,20 +5,54 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 02/08/2025
+## [0.3.2] - 08/08/2025
 
 ### Added
-- Styles
+- Dealer card flip animation
+
+## [0.3.1] - 07/08/2025
+
+### Changed
+- 1:1 payment.
+- Result message display.
+
+## [0.3.0] - 06/08/2025
+
+### Added
+- Styles.
+- Show profits or losses in result message.
+- Avoid text selection.
+
+### Changed
+- Color palette.
 
 ### Fixed
-- Wrong delay in stand action
+- Focus on bet input in mouse in event.
 
-## [0.2.0] - 24/07/2025
+## [0.2.0] - 03/08/2025
+
+### Added
+- Mouse and keyboard control support.
+- Clear hand animation.
+- Current player bet display.
+
+### Fixed
+- Enter key event listener for hit action.
+
+## [0.1.3] - 02/08/2025
+
+### Added
+- Styles.
+
+### Fixed
+- Wrong delay in stand action.
+
+## [0.1.2] - 24/07/2025
 
 ### Added
 - Delay between actions.
 
-## [0.1.0] - 23/07/2025
+## [0.1.1] - 23/07/2025
 
 ### Fixed
 - Dynamic ACE value, it could be 11 or 1.
@@ -32,10 +66,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flip hidden card to see it's value.
 - Player bet action.
 
-## [0.0.1] - 21/07/2025
+## [0.1.0] - 21/07/2025
 
 ### Added
-- First simple functional version.
+- First simple functional version with console logs.
 - Hit and stand actions.
 - Card drawing.
 - Score.